third_party/WebKit/LayoutTests/inspector/elements/animation-timeline-expected.txt

Tests the display of animations on the animation timeline.

>>>> Animation with start delay only
<svg class="animation-ui" height="80" width="682.67" style="margin-left: -8px; transform: translateX(0px);"><g style="transform: translateX(222.22px);"><g style="transform: translateX(0px);"><line class="animation-line" x1="8" y1="32" y2="32" x2="452.44" style="stroke: rgb(0, 0, 0);"></line><path class="animation-keyframe" d="M 0 32 L 0.02 13.35 L 0.94 13.35 L 3.51 13.35 L 7.66 13.35 L 13.28 13.35 L 20.28 13.35 L 28.59 13.35 L 38.10 13.35 L 48.73 13.35 L 60.39 13.35 L 72.99 13.35 L 86.44 13.35 L 100.65 13.35 L 115.54 13.35 L 131.00 13.35 L 146.95 13.35 L 163.31 13.35 L 179.98 13.35 L 196.87 13.35 L 213.90 13.35 L 230.96 13.35 L 247.99 13.35 L 264.88 13.35 L 281.54 13.35 L 297.89 13.35 L 313.83 13.35 L 329.28 13.35 L 344.15 13.35 L 358.34 13.35 L 371.77 13.35 L 384.35 13.35 L 395.99 13.35 L 406.60 13.35 L 416.08 13.35 L 424.36 13.35 L 431.33 13.35 L 436.92 13.35 L 441.03 13.35 L 443.56 13.35 L 444.44 13.35 L 444.44 32 Z" style="transform: translateX(8px); fill: rgb(0, 0, 0);"></path><circle class="animation-endpoint" cx="8.00" cy="32" r="4" style="stroke: rgb(0, 0, 0); fill: rgb(0, 0, 0);"></circle><circle class="animation-endpoint" cx="452.44" cy="32" r="4" style="stroke: rgb(0, 0, 0); fill: rgb(0, 0, 0);"></circle></g><g class="animation-tail-iterations"></g></g><line class="animation-delay-line" x1="8" y1="32" y2="32" x2="230.22" style="stroke: rgb(0, 0, 0);"></line><line class="animation-delay-line" x1="8" y1="32" y2="32" x2="8.00" style="stroke: rgb(0, 0, 0); transform: translateX(666.67px);"></line></svg>
>>>> Animation with start and end delay
<svg class="animation-ui" height="80" width="682.67" style="margin-left: -8px; transform: translateX(0px);"><g style="transform: translateX(3.28px);"><g style="transform: translateX(0px);"><line class="animation-line" x1="8" y1="32" y2="32" x2="664.81" style="stroke: rgb(0, 0, 0);"></line><path class="animation-keyframe" d="M 0 32 L 0.03 13.35 L 1.39 13.35 L 5.19 13.35 L 11.32 13.35 L 19.62 13.35 L 29.97 13.35 L 42.25 13.35 L 56.31 13.35 L 72.02 13.35 L 89.25 13.35 L 107.87 13.35 L 127.75 13.35 L 148.75 13.35 L 170.74 13.35 L 193.59 13.35 L 217.17 13.35 L 241.34 13.35 L 265.98 13.35 L 290.94 13.35 L 316.10 13.35 L 341.33 13.35 L 366.49 13.35 L 391.44 13.35 L 416.07 13.35 L 440.23 13.35 L 463.79 13.35 L 486.62 13.35 L 508.59 13.35 L 529.57 13.35 L 549.42 13.35 L 568.01 13.35 L 585.21 13.35 L 600.88 13.35 L 614.90 13.35 L 627.13 13.35 L 637.44 13.35 L 645.70 13.35 L 651.76 13.35 L 655.51 13.35 L 656.81 13.35 L 656.81 32 Z" style="transform: translateX(8px); fill: rgb(0, 0, 0);"></path><circle class="animation-endpoint" cx="8.00" cy="32" r="4" style="stroke: rgb(0, 0, 0); fill: rgb(0, 0, 0);"></circle><circle class="animation-endpoint" cx="664.81" cy="32" r="4" style="stroke: rgb(0, 0, 0); fill: rgb(0, 0, 0);"></circle></g><g class="animation-tail-iterations"></g></g><line class="animation-delay-line" x1="8" y1="32" y2="32" x2="11.28" style="stroke: rgb(0, 0, 0);"></line><line class="animation-delay-line" x1="8" y1="32" y2="32" x2="14.57" style="stroke: rgb(0, 0, 0); transform: translateX(660.1px);"></line></svg>
>>>> Animation canceled
<svg class="animation-ui animation-ui-canceled" height="80" width="682.67" style="margin-left: -8px; transform: translateX(0px);"><g style="transform: translateX(3.28px);"><g style="transform: translateX(0px);"><line class="animation-line" x1="8" y1="32" y2="32" x2="664.81" style="stroke: rgb(0, 0, 0);"></line><path class="animation-keyframe" d="M 0 32 L 0.03 13.35 L 1.39 13.35 L 5.19 13.35 L 11.32 13.35 L 19.62 13.35 L 29.97 13.35 L 42.25 13.35 L 56.31 13.35 L 72.02 13.35 L 89.25 13.35 L 107.87 13.35 L 127.75 13.35 L 148.75 13.35 L 170.74 13.35 L 193.59 13.35 L 217.17 13.35 L 241.34 13.35 L 265.98 13.35 L 290.94 13.35 L 316.10 13.35 L 341.33 13.35 L 366.49 13.35 L 391.44 13.35 L 416.07 13.35 L 440.23 13.35 L 463.79 13.35 L 486.62 13.35 L 508.59 13.35 L 529.57 13.35 L 549.42 13.35 L 568.01 13.35 L 585.21 13.35 L 600.88 13.35 L 614.90 13.35 L 627.13 13.35 L 637.44 13.35 L 645.70 13.35 L 651.76 13.35 L 655.51 13.35 L 656.81 13.35 L 656.81 32 Z" style="transform: translateX(8px); fill: rgb(0, 0, 0);"></path><circle class="animation-endpoint" cx="8.00" cy="32" r="4" style="stroke: rgb(0, 0, 0); fill: rgb(0, 0, 0);"></circle><circle class="animation-endpoint" cx="664.81" cy="32" r="4" style="stroke: rgb(0, 0, 0); fill: rgb(0, 0, 0);"></circle></g><g class="animation-tail-iterations"></g></g><line class="animation-delay-line" x1="8" y1="32" y2="32" x2="11.28" style="stroke: rgb(0, 0, 0);"></line><line class="animation-delay-line" x1="8" y1="32" y2="32" x2="14.57" style="stroke: rgb(0, 0, 0); transform: translateX(660.1px);"></line></svg>
